+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Post v1 a change is further considered breaking if the state file format is inco
 
 * Moved messages to external files using QuasiQuotes. ([#155](https://github.com/hjwylde/werewolf/issues/155))
 * Moved role names, descriptions and rules to external files using QuasiQuotes. ([#155](https://github.com/hjwylde/werewolf/issues/155))
+* Removed `--force` option from `end`. ([#224](https://github.com/hjwylde/werewolf/issues/224))
 
 ### v1.1.1.0